changed setup to only install pnpm node and cli, remove dependencies from root, made "vibe setup"
