+++ cli/src/cli.ts
@@ -3,6 +3,7 @@ import { Command } from "commander";
 import { runStart } from "./commands/start.ts";
 import { runTest } from "./commands/test.ts";
 import { runHelp } from "./commands/help.ts";
+import { runSetup } from "./commands/setup.ts";
 
 const program = new Command();
 
@@ -26,4 +27,9 @@ program
   .description("Run backend tests and check for failures")
   .action(runTest);
 
-program.parse(process.argv);
+program
+  .command("setup")
+  .description("Complete the setup of the project")
+  .action(runSetup);
+
+program.parse(process.argv)
