+++ cli/src/steps/firebase-emulators.ts
@@ -2,12 +2,9 @@
 import fs from "fs";
 import path from "path";
 import { execSync } from "child_process";
-import { fileURLToPath } from "url";
+import { findProjectRoot } from "../findRoot.ts";
 
-// Resolve __dirname for ES modules
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-const rootDir = path.resolve(__dirname, "../../");
+const rootDir = findProjectRoot();
 const backendDir = path.join(rootDir, "backend");
 const statePath = path.join(rootDir, ".vibe.json");
